argo-executor/pyproject.toml
[tool.poetry]
name = "magnus_extension_argo_executor"
version = "0.1.0"
description = "Self traversing compute in AWS batch"
authors = ["Vijay Vammi <vijay.vammi@astrazeneca.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/AstraZeneca/magnus-extensions/"
repository = "https://github.com/AstraZeneca/magnus-extensions/"
documentation = "https://astrazeneca.github.io/magnus-extensions/"

[tool.poetry.dependencies]
python = "^3.6.1"
magnus = "*"
docker = "^5.0.3"
pydantic = "^1.9.0"
"ruamel.yaml" = "*"



[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
autopep8 = "*"
pylint = "*"
mypy = "^0.931"
tox = "^3.24.5"

# Use this section to provide the namespace to extend and implementation
[tool.poetry.plugins."magnus.executor.BaseExecutor"]
"argo-workflow" = "magnus_extension_argo_executor.argo_executor:ArgoExecutor"

# Use this provide namespace of the integration
# naming convention: {executor.name}-{service}-{service.name}
#[tool.poetry.plugins."magnus.integration.BaseIntegration"]
#"aws-batch-traverse-run_log_store-file-system" = "magnus_extension_aws_batch_traverse.integration:LocalAWSBatchComputeRunLogStoreFileSystem"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
